requirements.txt
pillow~=10.2.0
tqdm~=4.66.1